Added static_cast<double>

+++ tasks/hypotenuse/hypotenuse.cpp
@@ -1,5 +1,5 @@
 #include "hypotenuse.h"
 
 double CalculateHypotenuse(int64_t a, int64_t b) {
-    return sqrt(a * a + b * b);
+    return sqrt(static_cast<double>(a * a + b * b));
 };
